% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDispersionsddsList.R
\name{calculateDispersionsddsList}
\alias{calculateDispersionsddsList}
\title{Calculate dispersions}
\usage{
calculateDispersionsddsList(ddsList, maxNForDisp)
}
\arguments{
\item{ddsList}{A list of DESeqDataSets}

\item{maxNForDisp}{If any data set contains more than \code{maxNForDisp}
samples, \code{maxNForDisp} of them will be randomly sampled before the
dispersions are calculated, in order to speed up calculations}
}
\value{
A list of the same length as the input list. Each element in the list
  is itself a list, containing a DGEList and a DESeqDataSet with calculated
  dispersions.
}
\description{
Calculate the dispersions for each data set in a list of DESeqDataSets, using
both edgeR and DESeq2.
}
\author{
Charlotte Soneson
}
\keyword{internal}
