% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCorrelations.R
\name{calculateSampleCorrs}
\alias{calculateSampleCorrs}
\title{Calculate Spearman correlation between sample pairs}
\usage{
calculateSampleCorrs(ddsList, maxNForCorr, seed = 123)
}
\arguments{
\item{ddsList}{List of lists, with one element per data set. Each element is
a list containing a DGEList and a DESeqDataSet, with calculated
dispersions (e.g., output from \code{\link{calculateDispersionsddsList}}).}

\item{maxNForCorr}{Maximal number of samples to use for correlation
calculation. If the number of samples in a data set exceeds
\code{maxNForCorr}, \code{maxNForCorr} samples will be randomly selected
for calculation of correlations.}

\item{seed}{The random seed used for downsampling/selection of samples.}
}
\value{
A data frame with pairwise sample correlations for each data set
}
\description{
This function is typically not called directly by the user, but is used when
generating the countsimQC report.
}
\examples{
data(countsimExample)
## Subset genes to speed up calculations
sub <- lapply(countsimExample, function(x) x[1:100, ])
## Calculate dispersions
obj <- calculateDispersionsddsList(ddsList = sub,
                                   maxNForDisp = 5, seed = 123)
## Calculate sample correlations
corrs <- calculateSampleCorrs(obj, maxNForCorr = 5, seed = 123)

}
\author{
Charlotte Soneson
}
