% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsim_example.R
\docType{data}
\name{countsim_example}
\alias{countsim_example}
\title{Example data set with three count data sets}
\format{A named list with three elements, each corresponding to a (real or simulated) data set.}
\usage{
countsim_example
}
\value{
A named list with three elements, each corresponding to a (real or simulated) data set.
}
\description{
A named list with three elements, each corresponding to a (real or simulated) data set.
}
\keyword{datasets}
