% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsimQC-pkg.R
\docType{package}
\name{countsimQC-package}
\alias{countsimQC}
\alias{countsimQC-package}
\title{countsimQC}
\description{
countsimQC
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoneson/countsimQC}
  \item Report bugs at \url{https://github.com/csoneson/countsimQC/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})

}
\keyword{internal}
