% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStats.R
\name{defaultStats}
\alias{defaultStats}
\title{Return a vector of NA scores}
\usage{
defaultStats(n, withP = FALSE)
}
\arguments{
\item{n}{Number of columns to use for the comparison}

\item{withP}{Whether or not to include p-value columns}
}
\value{
A vector with NA values for all applicable statistics
}
\description{
Return a vector of NA scores
}
\author{
Charlotte Soneson
}
\keyword{internal}
