% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_individual_plots.R
\name{generate_individual_plots}
\alias{generate_individual_plots}
\title{Generate individual plots from countsimQC_report output}
\usage{
generate_individual_plots(ggplots_rds, device = "png", output_dir = "./",
  n_datasets = 2)
}
\arguments{
\item{ggplots_rds}{The path to a .rds file generated by
\code{countsimQC_report} by setting \code{save_ggplots = TRUE}.}

\item{device}{One of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff",
"png", "bmp", "svg" or "wmf" (windows only) (will be provided to the
\code{ggsave} function from the \code{ggplot2} package).}

\item{output_dir}{The output directory where the plots will be generated.}

\item{n_datasets}{The number of data sets that are compared in the figures.
This is needed to set the size of the plots correctly.}
}
\value{
No value is returned, but plots are generated in the designated
  output directory.
}
\description{
Generate separate plots for all evaluation criteria using the collection of
ggplot objects that can be saved when generating a countsimQC report (by
setting \code{save_ggplots = TRUE}).
}
\author{
Charlotte Soneson
}
