% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(dds_list, show_code = FALSE, output_format = NULL,
  output_file = NULL, output_dir = "./", ...)
}
\arguments{
\item{dds_list}{Named list of DESeqDataSets to compare}

\item{show_code}{Whether or not to include code in report}

\item{output_format}{The output format of the report. If set to NULL, an html
report will be generated. Other options are "html_document" or
"pdf_document", or "all" to generate both an html report and a pdf report.}

\item{output_file}{The name of the report}

\item{output_dir}{The directory where the report will be saved}

\item{...}{Other arguments that will be passed to \code{rmarkdown::render}}
}
\value{
No value is returned, but a report is generated in the
  \code{output_dir} directory.
}
\description{
Generate report
}
\author{
Charlotte Soneson, using a template from Nicholas Hamilton
  (http://stackoverflow.com/questions/37097535/generate-report-in-r)
}
