% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCorrelations.R
\name{calculateFeatureCorrs}
\alias{calculateFeatureCorrs}
\title{Calculate Spearman correlation between variable pairs}
\usage{
calculateFeatureCorrs(ddsList, maxNForCorr, seed = 123)
}
\arguments{
\item{ddsList}{List of lists, with one element per data set. Each element is
a list containing a DGEList and a DESeqDataSet, with calculated
dispersions.}

\item{maxNForCorr}{If the number of variables in a data set exceeds
\code{maxNForCorr}, \code{maxNForCorr} variables will be randomly selected
for correlation calculation.}

\item{seed}{The seed for downsampling}
}
\value{
A data frame with pairwise variable correlations for each data set
}
\description{
Calculate Spearman correlation between variable pairs
}
\author{
Charlotte Soneson
}
