% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCorrelations.R
\name{calculateFeatureCorrs}
\alias{calculateFeatureCorrs}
\title{Calculate Spearman correlation between feature pairs}
\usage{
calculateFeatureCorrs(ddsList, maxNForCorr, seed = 123)
}
\arguments{
\item{ddsList}{List of lists, with one element per data set. Each element is
a list containing a DGEList and a DESeqDataSet, with calculated
dispersions (e.g., output from \code{\link{calculateDispersionsddsList}}).}

\item{maxNForCorr}{Maximal number of features to use for calculation of
correlations. If the number of features in a data set exceeds
\code{maxNForCorr}, \code{maxNForCorr} features will be randomly selected
for calculation of correlations.}

\item{seed}{The random seed used for downsampling/selection of features.}
}
\value{
A data frame with pairwise feature correlations for each data set
}
\description{
This function is typically not called directly by the user, but is used when
generating the countsimQC report.
}
\examples{
data(countsimExample)
## Subset genes to speed up calculations
sub <- lapply(countsimExample, function(x) x[1:100, ])
## Calculate dispersions
obj <- calculateDispersionsddsList(ddsList = sub,
                                   maxNForDisp = 5, seed = 123)
## Calculate feature correlations
corrs <- calculateFeatureCorrs(obj, maxNForCorr = 5, seed = 123)

}
\author{
Charlotte Soneson
}
