% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCorrelations.R
\name{calculateFeatureCorrs}
\alias{calculateFeatureCorrs}
\title{Calculate Spearman correlation between feature pairs}
\usage{
calculateFeatureCorrs(ddsList, maxNForCorr)
}
\arguments{
\item{ddsList}{List of lists, with one element per data set. Each element is
a list containing a DGEList and a DESeqDataSet, with calculated
dispersions (e.g., output from \code{\link{calculateDispersionsddsList}}).}

\item{maxNForCorr}{Maximal number of features to use for calculation of
correlations. If the number of features in a data set exceeds
\code{maxNForCorr}, \code{maxNForCorr} features will be randomly selected
for calculation of correlations.}
}
\value{
A data frame with pairwise feature correlations for each data set
}
\description{
Calculate Spearman correlation between feature pairs
}
\author{
Charlotte Soneson
}
\keyword{internal}
