% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCorrelations.R
\name{calculateSampleCorrs}
\alias{calculateSampleCorrs}
\title{Calculate Spearman correlation between sample pairs}
\usage{
calculateSampleCorrs(ddsList, maxNForCorr, seed = 123)
}
\arguments{
\item{ddsList}{List of lists, with one element per data set. Each element is
a list containing a DGEList and a DESeqDataSet, with calculated
dispersions.}

\item{maxNForCorr}{If the number of samples in a data set exceeds
\code{maxNForCorr}, \code{maxNForCorr} samples will be randomly selected
for correlation calculation.}

\item{seed}{The seed for downsampling}
}
\value{
A data frame with pairwise sample correlations for each data set
}
\description{
Calculate Spearman correlation between sample pairs
}
\author{
Charlotte Soneson
}
