% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCorrelations.R
\name{calculateSampleCorrs}
\alias{calculateSampleCorrs}
\title{Calculate Spearman correlation between sample pairs}
\usage{
calculateSampleCorrs(ddsList, maxNForCorr)
}
\arguments{
\item{ddsList}{List of lists, with one element per data set. Each element is
a list containing a DGEList and a DESeqDataSet, with calculated
dispersions (e.g., output from \code{\link{calculateDispersionsddsList}}).}

\item{maxNForCorr}{Maximal number of samples to use for correlation
calculation. If the number of samples in a data set exceeds
\code{maxNForCorr}, \code{maxNForCorr} samples will be randomly selected
for calculation of correlations.}
}
\value{
A data frame with pairwise sample correlations for each data set
}
\description{
Calculate Spearman correlation between sample pairs
}
\author{
Charlotte Soneson
}
\keyword{internal}
