% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsimExample.R
\docType{data}
\name{countsimExample}
\alias{countsimExample}
\title{Example list with three count data sets}
\format{A named list with three elements, each corresponding to a (real or
  simulated) count data set.}
\usage{
countsimExample
}
\value{
A named list with three elements, each corresponding to a (real or
  simulated) count data set.
}
\description{
A named list with three elements, each corresponding to a (real or simulated)
count data set.
}
\keyword{datasets}
