% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsimExample_dfmat.R
\docType{data}
\name{countsimExample_dfmat}
\alias{countsimExample_dfmat}
\title{Example list with three count data sets in different formats}
\format{A named list with three elements, each corresponding to a (real or
  simulated) count data set.}
\usage{
countsimExample_dfmat
}
\value{
A named list with three elements, each corresponding to a (real or
  simulated) count data set.
}
\description{
A named list with three elements, each corresponding to a (real or simulated)
count data set. One of them is provided as a DESeqDataset, one as a count
data frame and one as a count matrix.
}
\keyword{datasets}
